moved consignee Detail from the component to the page
